requirements.txt
aiogram==3.12.0
aiosqlite==0.20.0
cryptography==43.0.1
python-dotenv==1.0.1
requests==2.32.3